requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
langgraph>=0.2.0
ortools>=9.8.3296
python-dotenv>=1.0.0
httpx>=0.26.0
sqlmodel>=0.0.14

# Optional: explanation layer only (install openai for LLM explanations)
openai>=1.12.0
